Add code example from crate docs to readme, use readme in crate docs
In order to simplify documentation, this commit moves most of the
documentation directly into the README.md. Additionally, to avoid
duplication, this commit pulls in the README.md into the crate docs.
This allows one source of truth for documentation and a common
experience between viewing the crate on github, crates.io, and docs.rs.